+++ pccontrol/UIKeyboard.h
@@ -4,5 +4,6 @@
 #import <Foundation/Foundation.h>
 
 NSString* inputTextFromRawData(UInt8 *eventData, NSError **error);
+NSString* keyboardVisibleFromRawData(NSError **error);
 
 #endif